+++ containers/daap-presigned-url/CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.4] 2023-10-03
+
+### Changed
+
+- removed datebase and table querystring parameters
+- created date-product-name & table-name variables from event body pathParameters
+- created body variable from event.body json string 
+- replaced refferences to datebase with data-product-name
+- replaced refferences to table with table-name
+
 ## [1.2.3]
 
 ### Changed